inference/requirements.txt
fastapi>=0.78.0
uvicorn>=0.18.0
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
python-multipart>=0.0.5
pydantic>=1.9.0

